Refactor: Improve code and update string resources
- Updated `ReceiptData` default quantity to 1.
- Modified integer and float extension functions for zero checks.
- Renamed `ChooseFolderNameBottomSheet.kt`.
- Added new string resources and updated existing ones for clarity.
- Introduced `BackNavigationButton.kt` composable.
