+++ app/src/main/java/com/iliatokarev/receipt_splitter_app/main/basic/BasicExtentionFun.kt
@@ -6,11 +6,11 @@ import java.util.Date
 import java.util.Locale
 import kotlin.math.round
 
-fun Int.isZero(): Boolean = this == 0
-fun Int.isNotZero(): Boolean = this > 0
+fun Int.isNotZero(): Boolean = this != 0
+fun Int.isPositive(): Boolean = this > 0
 fun Int.isMoreThanOne(): Boolean = this >1
-fun Float.isZero(): Boolean = this == 0F
-fun Float.isNotZero(): Boolean = this > 0
+fun Float.isNotZero(): Boolean = this != 0F
+fun Float.isPositive(): Boolean = this >= 0
 fun Int?.getOrZero(): Int {
     return if (this == null || this == 0) 0 else this
 }
